langgraph-author-agent/requirements.txt
langgraph-cli[inmem]
tavily-python
deepagents
langchain-google-genai
langchain-mcp-adapters
langchain-anthropic
python-dotenv
